# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EduApp",
        "description": "Documentation Swagger de l'API d'authentification",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/voeuxexa": {
            "get": {
                "tags": [
                    "Voeux Examen"
                ],
                "summary": "Lister les vœux d'examen de l'enseignant connecté",
                "operationId": "f4d31d5848aa138533b2c2807e797112",
                "responses": {
                    "200": {
                        "description": "Liste des vœux d'examen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voeux-examen": {
            "post": {
                "tags": [
                    "Voeux Examen"
                ],
                "summary": "Ajouter un vœu de surveillance",
                "operationId": "990b6c56a4fe52ddd6677c7923a17774",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_creneau"
                                ],
                                "properties": {
                                    "code_creneau": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vœu ajouté"
                    },
                    "400": {
                        "description": "Charge dépassée ou vœu existant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voeuxexa/{code_creneau}": {
            "delete": {
                "tags": [
                    "Voeux Examen"
                ],
                "summary": "Supprimer un vœu de surveillance",
                "operationId": "34dd146053bfbc6e100e041cb15b43f4",
                "parameters": [
                    {
                        "name": "code_creneau",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vœu supprimé"
                    },
                    "404": {
                        "description": "Vœu introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/voeux-examen/bulk": {
            "post": {
                "tags": [
                    "Voeux Examen"
                ],
                "summary": "Remplacer tous les vœux de surveillance",
                "operationId": "8d00c13668f4877f7f6df96ccd91c62b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voeux": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "code_creneau": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vœux mis à jour"
                    },
                    "400": {
                        "description": "Charge dépassée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/enseignant/charge-surveillance": {
            "get": {
                "tags": [
                    "Voeux Examen"
                ],
                "summary": "Récupérer la charge de surveillance",
                "operationId": "19d39a577d45b33590b3380e7ca0a8be",
                "responses": {
                    "200": {
                        "description": "Charge retournée"
                    },
                    "404": {
                        "description": "Enseignant non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/creneaux": {
            "get": {
                "tags": [
                    "Créneaux"
                ],
                "summary": "Lister les créneaux d'examen",
                "operationId": "e524890e3cd1b0cfe50db8f77aefc29b",
                "responses": {
                    "200": {
                        "description": "Liste des créneaux"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}